Move unit tests next to the files their testing
