+++ src/__tests__/Settings.test.ts
@@ -1,0 +1,9 @@
+import Settings from '../Settings.svelte';
+
+describe('Settings component', () => {
+  it('renders correctly', () => {
+    const target = document.createElement('div');
+    new Settings({ target });
+    expect(target.innerHTML).toMatchSnapshot();
+  });
+});
