+++ src/__tests__/loader.test.ts
@@ -1,4 +1,4 @@
-// import loaderScript from '../src/loader.js';
+// import loaderScript from '../loader';
 
 describe('Loader script', () => {
   it.todo('should do something');
