+++ src/components/__tests__/BookmarkNode.test.ts
@@ -1,4 +1,4 @@
-import BookmarkNode from '../../src/components/BookmarkNode.svelte';
+import BookmarkNode from '../BookmarkNode.svelte';
 
 const _node = {};
