+++ src/components/__tests__/Search.test.ts
@@ -1,4 +1,4 @@
-import Search from '../../src/components/Search.svelte';
+import Search from '../Search.svelte';
 
 describe('Search component', () => {
   it('renders correctly', () => {
