--- test/components/BookmarkGroup.test.ts
@@ -1,9 +1,0 @@
-import BookmarkGroup from '../../src/components/BookmarkGroup.svelte';
-
-describe('BookmarkGroup component', () => {
-  it('renders correctly', () => {
-    const target = document.createElement('div');
-    new BookmarkGroup({ target });
-    expect(target.innerHTML).toMatchSnapshot();
-  });
-});
